+++ source/core/scene/shape/triangle/triangle_mesh_provider.cpp
@@ -335,9 +335,9 @@ Shape* Provider::load_binary(std::istream& stream, thread::Pool& thread_pool) no
     uint64_t json_size = 0;
     stream.read(reinterpret_cast<char*>(&json_size), sizeof(uint64_t));
 
-    memory::Array<char> json_string(json_size + 1);
+    memory::Array<char> json_string(uint32_t(json_size) + 1);
     stream.read(json_string.data(), std::streamsize(json_size * sizeof(char)));
-    json_string[json_size] = 0;
+    json_string[uint32_t(json_size)] = 0;
 
     std::string error;
     auto const  root = json::parse_insitu(json_string.data(), error);
